move getUnsupportedExtension to utils file
